renames fetch_closest_monitor to get_closest_monitor
