+++ deno.json
@@ -25,7 +25,7 @@
     "./collocation_sites/fetch_collocation_sites": "./lib/collocation_sites/fetch_collocation_sites.ts",
     "./monitors": "./lib/monitors/mod.ts",
     "./monitors/fetch_monitors": "./lib/monitors/fetch_monitors.ts",
-    "./monitors/fetch_closest_monitor": "./lib/monitors/fetch_closest_monitor.ts",
+    "./monitors/get_closest_monitor": "./lib/monitors/get_closest_monitor.ts",
     "./monitors/fetch_monitor_details": "./lib/monitors/fetch_monitor_details.ts",
     "./monitors/fetch_monitor_entries": "./lib/monitors/fetch_monitor_entries.ts",
     "./monitors/fetch_monitor_entries_csv": "./lib/monitors/fetch_monitor_entries_csv.ts",
