+++ lib/monitors/get_closest_monitor.ts
