+++ lib/monitors/mod.ts
@@ -1,4 +1,4 @@
-export * from "./fetch_closest_monitor.ts";
+export * from "./get_closest_monitor.ts";
 export * from "./fetch_monitor_details.ts";
 export * from "./fetch_monitor_entries.ts";
 export * from "./fetch_monitor_entries_csv.ts";
